Login and Signup pages
Fixes the inconsistent field size.

added toggle password visibility for both pages

+++ presentationLayer/react-app/src/pages/SignupPage.js
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
+import { FaEye, FaEyeSlash } from "react-icons/fa";
 
 import "../styles/Form.css";
-
 
 function SignupPage() {
 	let navigate = useNavigate();
@@ -10,37 +10,48 @@ function SignupPage() {
 	const [password, setPassword] = useState("");
 	const [confirmPassword, setConfirmPassword] = useState("");
 
+	const [showPassword, setShowPassword] = useState(false);
+	const [showConfirmedPassword, setShowConfirmedPassword] = useState(false);
+
+	const togglePasswordVisibility = () => {
+		setShowPassword(!showPassword);
+	}
+
+	const toggleConfirmedPasswordVisibility = () => {
+		setShowConfirmedPassword(!showConfirmedPassword);
+	}
+
 	const URL = localStorage.getItem("url") + "/signup";
 
 	const myHeaders = new Headers();
 	myHeaders.append("Content-Type", "application/json");
 
 	async function senddata() {
-        await fetch(URL, {
-            method: "POST",
-            headers: myHeaders,
-            body: JSON.stringify({
-                email: email,
-                password: password
-            }),
-        }).then(response => {
-			let result = response.json();
-			if (response.status>=200 && response.status<300){
-				localStorage.setItem("email", email);
-				localStorage.setItem("password", password);
-				navigate('/Homepage');
-				return result
-			}
-			else{
-				return result
-			}
-		}).then(data => {
+		await fetch(URL, {
+			method: "POST",
+			headers: myHeaders,
+			body: JSON.stringify({
+				email: email,
+				password: password,
+			}),
 		})
+			.then((response) => {
+				let result = response.json();
+				if (response.status >= 200 && response.status < 300) {
+					localStorage.setItem("email", email);
+					localStorage.setItem("password", password);
+					navigate("/Homepage");
+					return result;
+				} else {
+					return result;
+				}
+			})
+			.then((data) => {});
 		return false;
 	}
 
-    //state variable to check if the passwords are the same
-    let issame = password === confirmPassword
+	//state variable to check if the passwords are the same
+	let issame = password === confirmPassword;
 	return (
 		<div className="login-container">
 			<div className="login-card">
@@ -49,35 +60,63 @@ function SignupPage() {
 
 				<div className="input-group">
 					<label>Email</label>
-					<input 
-						type="email"
-						value = {email}
-						onChange = {(event) => setEmail(event.target.value)}
-						placeholder="Enter your email"
-					/>
+					<div className="credential-wrapper">
+						<input
+							type="email"
+							value={email}
+							onChange={(event) => setEmail(event.target.value)}
+							placeholder="Enter your email"
+						/>
+					</div>
 				</div>
 
 				<div className="input-group">
 					<label>Password</label>
-					<input 
-						type= "password"
-						value = {password}
-						onChange = {(event) => setPassword(event.target.value)}
-						placeholder="Create a password"
-					/>
+					<div className="credential-wrapper">
+						<input
+							type={showPassword ? "text" : "password"}
+							value={password}
+							onChange={(event) => setPassword(event.target.value)}
+							placeholder="Create a password"
+						/>
+						<button
+							type="button"
+							class="toggle-button"
+							onClick={togglePasswordVisibility}
+						>
+							{showPassword ? (
+								<FaEyeSlash className="show-password-icon" />
+							) : (
+								<FaEye className="show-password-icon" />
+							)}
+						</button>
+					</div>
 				</div>
 
 				<div className="input-group">
 					<label>Confirm Password</label>
-					<input
-						type = "password"
-						value = {confirmPassword}
-						onChange = {(event) => setConfirmPassword(event.target.value)}
-						placeholder = "Confirm password"
-					/>
+					<div className="credential-wrapper">
+						<input
+							type={showConfirmedPassword ? "text" : "password"}
+							value={confirmPassword}
+							onChange={(event) => setConfirmPassword(event.target.value)}
+							placeholder="Confirm password"
+						/>
+						<button
+							type = "button"
+							class = "toggle-button"
+							onClick={toggleConfirmedPasswordVisibility}
+						>
+							{showConfirmedPassword ? (
+								<FaEyeSlash className="show-password-icon"/>
+							) : (
+								<FaEye className="show-password-icon"/>
+							)}
+						</button>
+					</div>
 				</div>
 
-				<button 
+				<button
 					className="signin-button"
 					disabled={!issame}
 					type="submit"
@@ -92,7 +131,6 @@ function SignupPage() {
 			</div>
 		</div>
 	);
-
 }
 
 export default SignupPage;
